python/requirements-dev.txt
pytest
python-dotenv
openai
requests
websockets>=11
